requirements.txt
# FastAPI and dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic[email]==2.5.2

# Supabase
supabase==2.0.3
gotrue>=1.3.0,<2.0.0

# Utilities
python-dotenv==1.0.0